expression, plan: remove typeinfer.go since we infer all types in expression rewrite phrase
